refactor: Export `WindowInputs`
